Control the labels in different ways
Realize the control of the size related to the size of the words, color,
opacity of the rectangle labels and the color, boldness of the text.

Now, the rectangular label and the line contacted to it are in a class,
I was trying to put the draw words function in that class too, but
there's some use relationship among these functions, I don't know how to
change. Whatever, it works.